requirements.txt
pyttsx3
elevenlabs
python-dotenv
openai
pydub
google-cloud-aiplatform
google-cloud-texttospeech
pydub>=0.23.0
requests
python-dotenv
elevenlabs
ffmpeg-python
openai
moviepy==1.0.3
numpy==1.24.3
imageio>=2.5.0
decorator>=4.3.0
tqdm>=4.0.0
Pillow==9.5.0
scipy>=1.3.0
audiofile>=0.0.0
opencv-python>=4.5
colorlog
secure-smtplib
requests
python-dotenv
celery
redis
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
sse-starlette
